Lift logic to purge events controller
